remove redis

+++ Gateway/Models/APIDevice.cs
@@ -11,8 +11,8 @@ namespace Gateway.Models
         public string Name { get; set; }
         public string Ip { get; set; }
         public int RoomId { get; set; }
-        public float temprature { get; set; }
-        public float humidity { get; set; }
+        public double temprature { get; set; }
+        public double humidity { get; set; }
         public bool power { get; set; }
 
     }
